moved all track enums into track-enums file updated files to reference the new location
